rename tensorboard logger
