+++ theseus/utilities/loggers/__init__.py
@@ -1,5 +1,5 @@
 from .observer import LoggerObserver
-from .tf_logger import TensorboardLogger
+from .tsb_logger import TensorboardLogger
 from .image_writer import ImageWriter
 from .stdout_logger import StdoutLogger, FileLogger
 from .wandb_logger import WandbLogger
